backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-role-filters==1.0.1
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
unicodecsv==0.14.1
